Функции views.py, связанные со скачиванием, вынесены в отдельный файл
