+++ templates/account/statistics/download_modal.html
@@ -7,59 +7,62 @@
                 <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Закрыть"></button>
             </div>
 
-            <div class="modal-body">
-                <div class="row">
-                    <div class="col">
-                        <p>Какую информацию Вы хотите скачать?</p>
-                        <div class="row">
-                            <div class="col-4 d-grid">
-                                <input type="radio" class="btn-check" name="radio-reporttype" id="city-radio" autocomplete="off" checked>
-                                <label class="btn btn-outline-primary" for="city-radio">Города</label>
-                            </div>
-                            <div class="col-4 d-grid">
-                                <input type="radio" class="btn-check" name="radio-reporttype" id="region-radio" autocomplete="off">
-                                <label class="btn btn-outline-primary" for="region-radio">Регионы</label>
-                            </div>
-                            <div class="col-4 d-grid">
-                                <input type="radio" class="btn-check" name="radio-reporttype" id="area-radio" autocomplete="off">
-                                <label class="btn btn-outline-primary" for="area-radio">Округа</label></div>
-                            </div>
+            <form action="{% url 'download' %}" method="POST" id="download_form">
+                {% csrf_token %}
+                <div class="modal-body">
+                    <div class="row">
+                        <div class="col">
+                            <p>Какую информацию Вы хотите скачать?</p>
+                            <div class="row">
+                                <div class="col-4 d-grid">
+                                    <input type="radio" class="btn-check" name="radio-reporttype" id="city-radio" autocomplete="off" checked>
+                                    <label class="btn btn-outline-primary" for="city-radio">Города</label>
+                                </div>
+                                <div class="col-4 d-grid">
+                                    <input type="radio" class="btn-check" name="radio-reporttype" id="region-radio" autocomplete="off">
+                                    <label class="btn btn-outline-primary" for="region-radio">Регионы</label>
+                                </div>
+                                <div class="col-4 d-grid">
+                                    <input type="radio" class="btn-check" name="radio-reporttype" id="area-radio" autocomplete="off">
+                                    <label class="btn btn-outline-primary" for="area-radio">Округа</label></div>
+                                </div>
+                        </div>
+                    </div>
+    
+                    <div class="row mt-4">
+                        <div class="col">
+                            <p>В каком формате?</p>
+                            <div class="row">
+                                <div class="col-4 d-grid">
+                                    <input type="radio" class="btn-check" name="radio-filetype" id="txt-radio" autocomplete="off" checked>
+                                    <label class="btn btn-outline-primary" for="txt-radio">
+                                        <i class="fa-regular fa-file-lines"></i>&nbsp;&nbsp;&nbsp;TXT
+                                    </label>
+                                </div>
+                                <div class="col-4 d-grid">
+                                    <input type="radio" class="btn-check" name="radio-filetype" id="csv-radio" autocomplete="off">
+                                    <label class="btn btn-outline-primary" for="csv-radio">
+                                        <i class="fa-solid fa-file-csv"></i>&nbsp;&nbsp;&nbsp;CSV
+                                    </label>
+                                </div>
+                                <div class="col-4 d-grid">
+                                    <input type="radio" class="btn-check" name="radio-filetype" id="xls-radio" autocomplete="off">
+                                    <label class="btn btn-outline-primary" for="xls-radio">
+                                        <i class="fa-regular fa-file-excel"></i>&nbsp;&nbsp;&nbsp;XLS
+                                    </label></div>
+                                </div>
+                        </div>
+                    </div>
+    
+                    <hr>
+    
+                    <div class="row">
+                        <div class="col d-grid">
+                            <button type="submit" class="btn btn-success">Загрузить</button>
+                        </div>
                     </div>
                 </div>
-
-                <div class="row mt-4">
-                    <div class="col">
-                        <p>В каком формате?</p>
-                        <div class="row">
-                            <div class="col-4 d-grid">
-                                <input type="radio" class="btn-check" name="radio-filetype" id="txt-radio" autocomplete="off" checked>
-                                <label class="btn btn-outline-primary" for="txt-radio">
-                                    <i class="fa-regular fa-file-lines"></i>&nbsp;&nbsp;&nbsp;TXT
-                                </label>
-                            </div>
-                            <div class="col-4 d-grid">
-                                <input type="radio" class="btn-check" name="radio-filetype" id="csv-radio" autocomplete="off">
-                                <label class="btn btn-outline-primary" for="csv-radio">
-                                    <i class="fa-solid fa-file-csv"></i>&nbsp;&nbsp;&nbsp;CSV
-                                </label>
-                            </div>
-                            <div class="col-4 d-grid">
-                                <input type="radio" class="btn-check" name="radio-filetype" id="xls-radio" autocomplete="off">
-                                <label class="btn btn-outline-primary" for="xls-radio">
-                                    <i class="fa-regular fa-file-excel"></i>&nbsp;&nbsp;&nbsp;XLS
-                                </label></div>
-                            </div>
-                    </div>
-                </div>
-
-                <hr>
-
-                <div class="row">
-                    <div class="col d-grid">
-                        <button type="submit" class="btn btn-success">Загрузить</button>
-                    </div>
-                </div>
-            </div>
+            </form>
         </div>
     </div>
 </div>
